--- sedaro/src/sedaro/models/repository/RepositoryManager.py
@@ -1,10 +1,0 @@
-from typing import ClassVar
-
-from ..BaseModelManager import BaseModelManager
-from .Repository import Repository
-
-
-class RepositoryManager(BaseModelManager[Repository]):
-
-    _BASE_PATH: 'ClassVar[str]' = '/models/repositories'
-    _MODEL = Repository
