+++ sedaro/src/sedaro/models/repository/main.py
@@ -1,6 +1,7 @@
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, ClassVar
 
 from ..BaseModel import BaseModel
+from ..BaseModelManager import BaseModelManager
 
 if TYPE_CHECKING:
     from ..workspace.main import Workspace
@@ -14,3 +15,9 @@ class Repository(BaseModel):
         '''Get the workspace of the repository.'''
         from ..workspace.main import Workspace
         return self._get_rel('workspace', Workspace)
+
+
+class RepositoryManager(BaseModelManager[Repository]):
+
+    _BASE_PATH: 'ClassVar[str]' = '/models/repositories'
+    _MODEL = Repository
