+++ tests/test_model_managers.py
@@ -6,7 +6,7 @@ from config import API_KEY, HOST
 from sedaro import SedaroApiClient
 from sedaro.exceptions import SedaroApiException
 from sedaro.models.project.main import Project
-from sedaro.models.repository.Repository import Repository
+from sedaro.models.repository.main import Repository
 from sedaro.models.workspace.main import Workspace
 from sedaro.settings import SCENARIO
 
@@ -93,7 +93,7 @@ def test_repository():
         repo_new = sedaro.Repository.create(name='new repo', metamodelType=SCENARIO, workspace=w_new.id)
         assert isinstance(repo_new, Repository)
         assert repo_new.metamodelType == SCENARIO
-        assert repo_new.workspace == w_new.id
+        assert repo_new.workspace.id == w_new.id
         assert repo_new.name == 'new repo'
 
         repo_new2 = sedaro.Repository.get(repo_new.id)
